requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
requests==2.32.3
gradio==4.36.1
scikit-learn==1.5.0
kagglehub